+++ lib/rust-vc-utils/src/seq_util.rs
@@ -1,3 +1,23 @@
+pub fn rev_comp(dna: &[u8]) -> Vec<u8> {
+    let comp_base = |x: &u8| -> u8 {
+        match *x {
+            b'A' => b'T',
+            b'T' => b'A',
+            b'C' => b'G',
+            b'G' => b'C',
+            b'N' => b'N',
+            b'a' => b't',
+            b't' => b'a',
+            b'c' => b'g',
+            b'g' => b'c',
+            b'n' => b'n',
+            y => panic!("Unsupported DNA string character {y}"),
+        }
+    };
+
+    dna.iter().rev().map(comp_base).collect::<Vec<_>>()
+}
+
 pub fn rev_comp_in_place(dna: &mut [u8]) {
     let comp_base = |x: &mut u8| {
         *x = match *x {
@@ -5,7 +25,13 @@ pub fn rev_comp_in_place(dna: &mut [u8]) {
             b'T' => b'A',
             b'C' => b'G',
             b'G' => b'C',
-            _ => *x,
+            b'N' => b'N',
+            b'a' => b't',
+            b't' => b'a',
+            b'c' => b'g',
+            b'g' => b'c',
+            b'n' => b'n',
+            y => panic!("Unsupported DNA string character {y}"),
         };
     };
 
@@ -24,6 +50,14 @@ pub fn rev_comp_in_place(dna: &mut [u8]) {
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[test]
+    fn test_rev_comp() {
+        let input = b"NNATGCG".to_vec();
+        let expected_output = b"CGCATNN".to_vec();
+        let output = rev_comp(&input);
+        assert_eq!(output, expected_output);
+    }
 
     #[test]
     fn test_rev_comp_in_place() {